[#6] feat: 클러스터링 연결 및 폴더 구조에 맞게 파일 위치 변경
