+++ Chac/Chac/Source/Domain/Interfaces/ClusteringStrategy.swift
